Don't incorrectly add finalizer to connection (!!!)
This fixes mysterious sporadic disconnects. The bug was that
addFinalizer adds finalizer to a thunk, not to the evaluated `conn` and
then fires on the wrong moment. This could be fixed by e.g.
`evaluate conn >>= flip addFinalizer …`
but I didn't add it as auto-closing is not very expected (XXX?).